move artifact utils tests out of test_utils
