+++ tests/tracking/test_utils.py
@@ -11,7 +11,6 @@ from mlflow.store.sqlalchemy_store import SqlAlchemyStore
 from mlflow.tracking.utils import _get_store, _TRACKING_URI_ENV_VAR, _TRACKING_USERNAME_ENV_VAR, \
     _TRACKING_PASSWORD_ENV_VAR, _TRACKING_TOKEN_ENV_VAR, _TRACKING_INSECURE_TLS_ENV_VAR, \
     get_db_profile_from_uri, TrackingStoreRegistry
-from mlflow.tracking.artifact_utils import _download_artifact_from_uri
 
 
 def test_get_store_file_store(tmp_wkdir):
@@ -280,46 +279,3 @@ def test_get_store_for_unregistered_scheme():
 
 def test_get_db_profile_from_uri_casing():
     assert get_db_profile_from_uri('databricks://aAbB') == 'aAbB'
-
-
-def test_artifact_can_be_downloaded_from_absolute_uri_successfully(tmpdir):
-    artifact_file_name = "artifact.txt"
-    artifact_text = "Sample artifact text"
-    local_artifact_path = tmpdir.join(artifact_file_name).strpath
-    with open(local_artifact_path, "w") as out:
-        out.write(artifact_text)
-
-    logged_artifact_path = "artifact"
-    with mlflow.start_run():
-        mlflow.log_artifact(local_path=local_artifact_path, artifact_path=logged_artifact_path)
-        artifact_uri = mlflow.get_artifact_uri(artifact_path=logged_artifact_path)
-
-    downloaded_artifact_path = os.path.join(
-        _download_artifact_from_uri(artifact_uri), artifact_file_name)
-    assert downloaded_artifact_path != local_artifact_path
-    assert downloaded_artifact_path != logged_artifact_path
-    with open(downloaded_artifact_path, "r") as f:
-        assert f.read() == artifact_text
-
-
-def test_download_artifact_from_absolute_uri_persists_data_to_specified_output_directory(tmpdir):
-    artifact_file_name = "artifact.txt"
-    artifact_text = "Sample artifact text"
-    local_artifact_path = tmpdir.join(artifact_file_name).strpath
-    with open(local_artifact_path, "w") as out:
-        out.write(artifact_text)
-
-    logged_artifact_subdir = "logged_artifact"
-    with mlflow.start_run():
-        mlflow.log_artifact(local_path=local_artifact_path, artifact_path=logged_artifact_subdir)
-        artifact_uri = mlflow.get_artifact_uri(artifact_path=logged_artifact_subdir)
-
-    artifact_output_path = tmpdir.join("artifact_output").strpath
-    os.makedirs(artifact_output_path)
-    _download_artifact_from_uri(artifact_uri=artifact_uri, output_path=artifact_output_path)
-    assert logged_artifact_subdir in os.listdir(artifact_output_path)
-    assert artifact_file_name in os.listdir(
-        os.path.join(artifact_output_path, logged_artifact_subdir))
-    with open(os.path.join(
-            artifact_output_path, logged_artifact_subdir, artifact_file_name), "r") as f:
-        assert f.read() == artifact_text
